Consolidate pre-compiled specializations into single `libraft` binary (#1333)
Now that we no longer need to rely on the FAISS dependency, this PR:
1. Consolidates  the `libraft-distance` and `libraft-nn` targets into a single `libraft` artifact
2. Introduces a new `raft::compiled` target to link against the new `libraft.so` binary. Removes `raft::distance`and `raft::nn` targets.
3. Consolidates `RAFT_DISTANCE_COMPILED` and `RAFT_NN_COMPILED` pre-processor vars into a single `RAFT_COMPILED` (to match similar pattern implementated by `spdlog`)
4. Consolidates `specializations.cuh` headers
5. Updates all docs, scripts, and build infra

This change has been a long time coming and is intended to be a 23.04 feature. This is further going to require updates to several projects downstream. Here's a checklist to track that progress:
- [x] offline announcement
- [x] cuml https://github.com/rapidsai/cuml/pull/5272
- [x] cugraph https://github.com/rapidsai/cugraph/pull/3348
- [x] cuopt https://github.com/rapidsai/cuopt/pull/1023
- [x] cugraph-ops https://github.com/rapidsai/cugraph-ops/pull/429


This PR depended on https://github.com/rapidsai/raft/pull/1340 (removing FAISS from the build) and on https://github.com/rapidsai/raft/pull/1202 (replacing the FAISS bfknn w/ our own), both of which have been merged. 

Closes #824

Authors:
  - Corey J. Nolet (https://github.com/cjnolet)
  - Ben Frederickson (https://github.com/benfred)

Approvers:
  - Sevag H (https://github.com/sevagh)
  - Divye Gala (https://github.com/divyegala)
  - Ben Frederickson (https://github.com/benfred)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/raft/pull/1333

+++ cpp/src/random/rmat_rectangular_generator_int_double.cu
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, NVIDIA CORPORATION.
+ * Copyright (c) 2022-2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,6 +18,6 @@
 
 namespace raft::runtime::random {
 
-FUNC_DEF(int64_t, double);
+FUNC_DEF(int, double);
 
 }  // namespace raft::runtime::random
